Build Custom Hook useDetails
